Add nix-install-url; lock to Nix 2.13.3
Previously, our action would spuriously break because Nix minor version
bumps are apparently as unstable as a major version bump.

This commit locks it to 2.13.3 unless there's a major change. The user
can override this using the nix-install-url action key, but doing so
will likely break the core script like it did previously.